==> 1.a. data_collection.ipynb <==
import pandas as pd

df1 = pd.read_csv("data/source/Bangalore.csv")

df1.head()

df1.shape

df1.columns

df1.dtypes

df1.describe()

ROUTE = "data/source/"

pd1 = pd.read_csv(ROUTE + "Bangalore.csv")

pd1.insert(1,"city", "Bangalore")

pd2 = pd.read_csv(ROUTE + "Chennai.csv")
pd2.insert(1,"city", "Chennai")

pd_combined = pd.concat([pd1,pd2], ignore_index=True)

pd_combined.head()

import os

dirs_files = [os.path.join(ROUTE, f) for f in os.listdir(ROUTE) if f.endswith('.csv')]

dirs_files

dfs = []
for file in dirs_files:
    df = pd.read_csv(file)
    city_name_1 = file.split('/')[-1]
    city_name = city_name_1.split('.csv')[0]
    df.insert(1,"city",city_name)
    dfs.append(df)


[df.shape for df in dfs]

df_combined = pd.concat(dfs, ignore_index=True)

df_combined.shape

df_combined = df_combined.sample(frac=1)

df_combined.head()

df_combined.to_csv("data/preprocessed/combine_data.csv", index=False)

df = pd.read_csv("data/preprocessed/combine_data.csv")
df

from sklearn.model_selection import train_test_split
train_df, test_df = train_test_split(df_combined,test_size=0.3, random_state=42)

test_df.shape

train_df.to_csv("data/preprocessed/train_data.csv")
test_df.to_csv("data/preprocessed/test_data.csv")
