==> 1.b. data_collection_refactored.ipynb <==
from utils import collect as col 

ROUTE = "data/source/"
df_combined, df_train, df_test = col.collect_and_join(ROUTE)

DEST_ROUTE="data/preprocessed/"
df_combined.to_csv(DEST_ROUTE + "combine_data.csv", index=False)
df_train.to_csv(DEST_ROUTE + "train_data.csv", index=False)
df_test.to_csv(DEST_ROUTE + "test_data.csv", index=False)
