==> 2. base_model.ipynb <==
import pandas as pd
import numpy as np

TRAIN_DATA = "data/preprocessed/train_data.csv"

train_df = pd.read_csv(TRAIN_DATA)

train_df.shape

train_df.isnull().sum()

train_df.head()

train_df['price_per_foot^2'] = train_df['Price']/train_df['Area']

train_df.head()

df_summary = train_df.groupby(['city','Location'])[['price_per_foot^2']].mean()

df_summary.shape

train_df['Location'].nunique()

train_df.shape

# creamos un var donde eliminamos espacios y digitos y pasamos a minusculas location
train_df['Standard_Location'] = train_df['Location'].str.lower().str.replace(' ','').str.replace(r'\d+','', regex=True)

train_df['Standard_Location'] = train_df['Standard_Location'].str.replace('sector','')

train_df['Standard_Location'].nunique()

df_summary_2 = train_df.groupby(['city','Standard_Location'])[['price_per_foot^2']].mean()

city_location_dict = df_summary_2.to_dict()

city_location_dict = city_location_dict['price_per_foot^2']

city_location_dict['Bangalore', 'abbigere']

def predict_base_model(data):
    result = data.apply(
        func=lambda x: city_location_dict[x['city'], x['Standard_Location']]*x['Area'], 
        axis=1)
    return result.values

def extend_features(data):
    data.loc[:,'Standard_Location'] = ( data['Location'].
                                       str.lower().
                                       str.replace(' ','').
                                       str.replace(r'\d+','', regex=True).
                                       str.replace('sector','') )
    return data

train_df_eval = train_df.iloc[:2]
train_df_eval_extent = extend_features(train_df.iloc[:2])
predict_base_model(train_df_eval_extent)

train_df_eval['Price']

TEST_DATA = "data/preprocessed/test_data.csv"
test_data = pd.read_csv(TEST_DATA)
test_data_extent = extend_features(test_data)
predict_base_model(test_data_extent)


df_summary_city = train_df.groupby(['city'])[['price_per_foot^2']].mean()

df_summary_city

city_average_dict = df_summary_city.to_dict()['price_per_foot^2']

city_average_dict

def predict_robust_base_model(data):
    def calculate_price(x):
        city_average = city_average_dict[x['city']]
        p_p_f = city_location_dict.get((x['city'], x['Standard_Location']),city_average)
        return p_p_f*x['Area']
    
    result = data.apply(
        func=calculate_price, 
        axis=1)
    return result.values

predict_robust_base_model(test_data_extent)

from sklearn.metrics import ( root_mean_squared_error, 
                             mean_absolute_error, 
                             mean_absolute_percentage_error )

y_true = test_data['Price'].values
y_pred = predict_robust_base_model(test_data_extent)
rmse = root_mean_squared_error(y_true, y_pred)
mae = mean_absolute_error(y_true, y_pred)
mape = mean_absolute_percentage_error(y_true, y_pred)
print(f"rmse: {rmse}")
print(f"mae: {mae}")
print(f"mape: {mape}")
