==> 4.b. feature_selection.ipynb <==
import pandas as pd
import numpy as  np

# Cargar nuestro datos postproccesados de entrenamiento.
X_EXPANDED_TEST = pd.read_pickle("data/postproccesed/X_EXPANDED.pkl")
y_train = pd.read_pickle("data/postproccesed/y.pkl")

X_EXPANDED_TEST.shape

from sklearn.feature_selection import VarianceThreshold



# Eliminamos datos con muy poca varianza
lowVarianceFilter = VarianceThreshold(0.05)
lowVarianceFilter.fit(X_EXPANDED_TEST)

X_wo_low_variance = lowVarianceFilter.transform(X_EXPANDED_TEST)

X_wo_low_variance.shape

X_wo_low_variancelow_df = pd.DataFrame(data=X_wo_low_variance,
                                        columns=lowVarianceFilter.get_feature_names_out(),
                                        index=X_EXPANDED_TEST.index)

# Eliminamos columnas repetidas. 
X_wo_low_variancelow_df_2 = X_wo_low_variancelow_df.loc[:,~X_wo_low_variancelow_df.columns.duplicated()].copy()

X_wo_low_variancelow_df_2.shape

from feature_engine.selection import DropCorrelatedFeatures, DropDuplicateFeatures


# Eliminamos duplicados.
filter_duplicates = DropDuplicateFeatures()
filter_duplicates.fit(X_wo_low_variancelow_df_2)
X_wo_duplicates = filter_duplicates.transform(X_wo_low_variancelow_df_2)

X_wo_duplicates.shape

# Eliminamos variables muy correladas
correlated_filter = DropCorrelatedFeatures(threshold=0.9)

correlated_filter.fit(X_wo_duplicates)
X_wo_high_correlations = correlated_filter.transform(X_wo_duplicates)

X_wo_high_correlations.shape

# Aplicamos um filtro más avanzado: ProbeFeatureSelection
from  feature_engine.selection import ProbeFeatureSelection
from sklearn.linear_model import LinearRegression

advanced_filtered = ProbeFeatureSelection(estimator=LinearRegression(),
                                          scoring='neg_mean_absolute_percentage_error',
                                          n_probes=3,
                                          distribution="normal",)

advanced_filtered.fit(X_wo_high_correlations, y_train)
final_X = advanced_filtered.transform(X_wo_high_correlations)

final_X.shape

# Guardamos por si queremos procesar más adelante
final_X.to_pickle("data/filtered/X_filter_1.pkl")

# Vamos nuestro primer modelo.

from sklearn.ensemble import RandomForestRegressor

rf_model = RandomForestRegressor()
rf_model.fit(final_X,y_train)

rf_model.score(final_X,y_train)

rf_model.predict(final_X.iloc[:2])

# Vamos ahora a cargar nuestro datos de test y
# aplicarle todos los pasos de preprocesamiento y selección
# que le hemos aplicado a entrenamiento

import pickle
with open('data/postproccesed/preproccesing_price_Transformer.pkl', 'rb') as f:
    price_transformer = pickle.load(f)

with open('data/postproccesed/preproccesing_scaler_price.pkl', 'rb') as f:
    scalar_price = pickle.load(f)


def unscale(model, data_points):
    prediction = model.predict(data_points)
    unscaled = scalar_price.inverse_transform(prediction.reshape(-1, 1))
    unnormed = price_transformer.inverse_transform(unscaled.reshape(-1, 1))
    return unnormed

unscale(rf_model, final_X.iloc[:2])

# cogemos el set de entrenamiento original
test_data_df = pd.read_csv("data/preprocessed/test_data.csv")

test_data_df.shape

test_data_df = test_data_df.replace({9:np.nan})
test_data_df[test_data_df.columns[5:]] = test_data_df[test_data_df.columns[5:]].replace({0:'NO', 1:'SI', np.nan:'NO_DISPONIBLE'})

test_data_df.head()

with open('data/postproccesed/preproccesing_beds_imputer.pkl', 'rb') as f:
    beds_imputer = pickle.load(f)

test_data_df['No. of Bedrooms'] = beds_imputer.transform(test_data_df[['No. of Bedrooms']])

with open('data/postproccesed/preproccesing_ohEncoder.pkl', 'rb') as f:
    oh_encoder = pickle.load(f)

cat_columns = ["city"] + test_data_df.columns[5:].to_list()
cat_columns
cat_oh_data = oh_encoder.transform(test_data_df[cat_columns])

df_cat_oh_data = pd.DataFrame(data=cat_oh_data, columns=oh_encoder.get_feature_names_out(), index=test_data_df.index)

with open('data/postproccesed/preproccesing_gapEncoder.pkl', 'rb') as f:
    gap_encoder = pickle.load(f)

cat_location_data = gap_encoder.transform(test_data_df['Location'])

with open('data/postproccesed/preproccesing_area_Transformer.pkl', 'rb') as f:
    area_transformer = pickle.load(f)

with open('data/postproccesed/preproccesing_beds_Transformer.pkl', 'rb') as f:
    bed_transformer = pickle.load(f)

with open('data/postproccesed/preproccesing_price_Transformer.pkl', 'rb') as f:
    price_transformer = pickle.load(f)

with open('data/postproccesed/preproccesing_scaler_area.pkl', 'rb') as f:
    area_scaler = pickle.load(f)

with open('data/postproccesed/preproccesing_scaler_beds.pkl', 'rb') as f:
    beds_scaler = pickle.load(f)

with open('data/postproccesed/preproccesing_scaler_price.pkl', 'rb') as f:
   price_scaler = pickle.load(f)

price_normal = price_transformer.transform(test_data_df[['Price']])
area_normal = area_transformer.transform(test_data_df[['Area']])
beds_normal = bed_transformer.transform(test_data_df[['No. of Bedrooms']])

price_normal_scaled = price_scaler.transform(price_normal)
area_normal_scaled = area_scaler.transform(area_normal)
beds_normal_scaled = beds_scaler.transform(beds_normal)

df_normal_scaled = pd.DataFrame(data={'price_normal_scaled':price_normal_scaled.flatten(), 
                                      'area_normal_scaled':area_normal_scaled.flatten(), 
                                      'beds_normal_scaled':beds_normal_scaled.flatten()}, 
                                      index=test_data_df.index)

with open('data/postproccesed/preproccesing_polyfeatures.pkl', 'rb') as f:
   polyfeatures = pickle.load(f)

features_to_cross = pd.concat([df_cat_oh_data, df_normal_scaled[['area_normal_scaled', 'beds_normal_scaled']]], axis=1)

crossed_features = polyfeatures.transform(features_to_cross)

df_crossed_features = pd.DataFrame(data=crossed_features, columns=polyfeatures.get_feature_names_out(), index=test_data_df.index)

X_EXPANDED = pd.concat([df_normal_scaled[['area_normal_scaled', 'beds_normal_scaled']],
                        df_cat_oh_data,
                        cat_location_data,
                        df_crossed_features], axis=1)

# antes de aplicar el filtrado de entrenamiento podemos guardar en postprocessed como test:
X_EXPANDED.to_pickle("data/postproccesed/X_EXPANDED_TEST.pkl")


test_X_wo_low_variance = lowVarianceFilter.transform(X_EXPANDED)

test_X_wo_low_variancelow_df = pd.DataFrame(data=test_X_wo_low_variance,
                                        columns=lowVarianceFilter.get_feature_names_out(),
                                        index=X_EXPANDED.index)

test_X_wo_low_variancelow_df_2 = test_X_wo_low_variancelow_df.loc[:,~test_X_wo_low_variancelow_df.columns.duplicated()].copy()

test_X_wo_duplicates = filter_duplicates.transform(test_X_wo_low_variancelow_df_2)

test_X_wo_high_correlations = correlated_filter.transform(test_X_wo_duplicates)

test_final_X = advanced_filtered.transform(test_X_wo_high_correlations)

# podemos guardarlo en filtered al igual que entrenamiento
test_final_X.to_pickle("data/filtered/X_test_filter_1.pkl")

test_final_X.shape

y_test = test_data_df['Price']
y_test.to_pickle("data/postproccesed/y_test.pkl")

y_test.shape

test_predictions = unscale(rf_model,test_final_X)

from sklearn.metrics import ( root_mean_squared_error, 
                             mean_absolute_error, 
                             mean_absolute_percentage_error )



y_true = y_test.values
y_pred = test_predictions
rmse = root_mean_squared_error(y_true, y_pred)
mae = mean_absolute_error(y_true, y_pred)
mape = mean_absolute_percentage_error(y_true, y_pred)
print(f"rmse: {rmse}")
print(f"mae: {mae}")
print(f"mape: {mape}")
