==> 5.b model_comparison_02.ipynb <==
import pandas as pd
import numpy as np

from sklearn.metrics import ( root_mean_squared_error, 
                             mean_absolute_error, 
                             mean_absolute_percentage_error )

DIR_X = "data/filtered/"
DIR_Y = "data/postproccesed/"
# leemos x de fichero
X_train = pd.read_pickle(DIR_X + "X_filter_1.pkl")
X_test = pd.read_pickle(DIR_X + "X_test_filter_1.pkl")
# leemos y de fichero
y_train = pd.read_pickle(DIR_Y + "y.pkl")
y_test = pd.read_pickle(DIR_Y + "y_test.pkl")

print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

## function to unscale price to original values.
import pickle
with open('data/postproccesed/preproccesing_price_Transformer.pkl', 'rb') as f:
    price_transformer = pickle.load(f)
with open('data/postproccesed/preproccesing_scaler_price.pkl', 'rb') as f:
    scalar_price = pickle.load(f)
def unscale(data_points):
    unscaled = scalar_price.inverse_transform(data_points.reshape(-1, 1))
    unnormed = price_transformer.inverse_transform(unscaled.reshape(-1, 1))
    return unnormed

from sklearn.ensemble import HistGradientBoostingRegressor
hb_m_1 = HistGradientBoostingRegressor(l2_regularization=0.5)
hb_m_1.fit(X_train, y_train)

predictions = hb_m_1.predict(X_test)
predictions_unscaled = unscale(predictions)

y_true = y_test.values
y_pred = predictions_unscaled
rmse = root_mean_squared_error(y_true, y_pred)
mae = mean_absolute_error(y_true, y_pred)
mape = mean_absolute_percentage_error(y_true, y_pred)
print(f"rmse: {rmse}")
print(f"mae: {mae}")
print(f"mape: {mape}")

from sklearn.linear_model import Ridge
rg = Ridge()
rg.fit(X_train, y_train)

predictions = rg.predict(X_test)
predictions_unscaled = unscale(predictions)

y_true = y_test.values
y_pred = predictions_unscaled
rmse = root_mean_squared_error(y_true, y_pred)
mae = mean_absolute_error(y_true, y_pred)
mape = mean_absolute_percentage_error(y_true, y_pred)
print(f"rmse: {rmse}")
print(f"mae: {mae}")
print(f"mape: {mape}")

from xgboost import XGBRegressor

xgb = XGBRegressor(n_estimators=10)
# fit model
xgb.fit(X_train, y_train)


predictions = xgb.predict(X_test)
predictions_unscaled = unscale(predictions)

y_true = y_test.values
y_pred = predictions_unscaled
rmse = root_mean_squared_error(y_true, y_pred)
mae = mean_absolute_error(y_true, y_pred)
mape = mean_absolute_percentage_error(y_true, y_pred)
print(f"rmse: {rmse}")
print(f"mae: {mae}")
print(f"mape: {mape}")
